third_party/WebKit/LayoutTests/images/zoomed-offset-size-expected.txt

This is a testharness.js-based test.
FAIL Image should return a size of 250x166 regardless of zoom level. assert_equals: expected 250 but got 251
Harness: the test ran to completion.
